working at the geostatistical theory perspective

+++ DeepLearning/src/main/scala/com/mai/scalaML/probabilities.scala
@@ -4,6 +4,6 @@ object TreeProb {
 
 val counts = Array (2,9,38,67,74,50,12,3)
 val sumCounts = counts.sum
-val probCounts = counts.map(i => i/sumCounts)
+val probCounts = counts.map(i  => i.toDouble/sumCounts)
 
 }
